EliteMobs 8.3.5 - [New] Added moveToTarget to script particles, allowing some particles to move towards or away from the center of a zone - [New] Added the MODIFY_DAMAGE script action, allowed scripts to modify damage - [New] Non-destructive improvements to the API surrounding listening to EliteMobs damage - [Tweak] Improved script error catching - [Tweak] Improved mount error catching - [Fix] Fixed important issue that would cause scripts that run for zones to not only affect one player inside of a zone - [Fix] Fixed default teleport for the under grove leading to the wrong lair - [Fix] Fixed error relate to quest tracking - [Fix] Fixed issue where procedurally generated quests were giving away exponentially more money than intended

+++ src/main/java/com/magmaguy/elitemobs/items/customloottable/CurrencyCustomLootEntry.java
@@ -41,7 +41,7 @@ public class CurrencyCustomLootEntry extends CustomLootEntry implements Serializ
                     break;
                 case "chance":
                     try {
-                        super.setChance(Integer.parseInt(strings[1]));
+                        super.setChance(Double.parseDouble(strings[1]));
                     } catch (Exception ex) {
                         errorMessage(rawString, configFilename, "chance");
                     }
